Created todaysprogress subview and put it on the analytics view

+++ Habito/Views/AnalyticsViews/BarGraphView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 import Charts
 
 struct BarGraphView: View {
-    let data: [Double]
+    let data: [Int]
     let labels: [String]
 
     var body: some View {
